Move ert._c_wrappers.job_queue to ert.job_queue
