+++ src/clib/lib/CMakeLists.txt
@@ -54,7 +54,7 @@ install(TARGETS _clib LIBRARY DESTINATION src/ert)
 # This extra deploy is only for ctest to be able to locate the script, this file
 # is otherwise deployed through setup.py.
 file(
-  COPY "${CMAKE_CURRENT_SOURCE_DIR}/../../ert/_c_wrappers/job_queue/qstat_proxy.sh"
+  COPY "${CMAKE_CURRENT_SOURCE_DIR}/../../ert/job_queue/qstat_proxy.sh"
   DESTINATION "${CMAKE_BINARY_DIR}/old_tests/job_queue"
   FILE_PERMISSIONS OWNER_EXECUTE OWNER_WRITE OWNER_READ)
