+++ tests/test_config_parsing/config_dict_generator.py
@@ -15,7 +15,7 @@ from pydantic import PositiveInt
 
 from ert._c_wrappers.enkf import ConfigKeys
 from ert._c_wrappers.enkf.config.field_config import TRANSFORM_FUNCTIONS
-from ert._c_wrappers.job_queue import QueueDriverEnum
+from ert.job_queue import QueueDriverEnum
 
 from .egrid_generator import EGrid, egrids
 from .observations_generator import Observation, observations
