+++ tests/unit_tests/c_wrappers/job_runner/test_jobmanager.py
@@ -8,8 +8,8 @@ import pytest
 from _ert_job_runner.reporting.message import Exited, Start
 from _ert_job_runner.runner import JobRunner
 from ert._c_wrappers.enkf import ErtConfig
-from ert._c_wrappers.job_queue import ExtJob
 from ert._c_wrappers.util import SubstitutionList
+from ert.job_queue import ExtJob
 
 # Test data generated by ForwardModel
 JSON_STRING = """
