+++ tests/unit_tests/c_wrappers/res/job_queue/test_job_queue_manager.py
@@ -7,7 +7,7 @@ from typing import Callable, List, TypedDict
 
 import pytest
 
-from ert._c_wrappers.job_queue import (
+from ert.job_queue import (
     Driver,
     JobQueue,
     JobQueueManager,
