+++ tests/unit_tests/c_wrappers/res/test_forward_model_json.py
@@ -10,8 +10,8 @@ from typing import List
 import pytest
 
 from ert._c_wrappers.enkf import ErtConfig
-from ert._c_wrappers.job_queue.ext_job import ExtJob
 from ert._c_wrappers.util import SubstitutionList
+from ert.job_queue import ExtJob
 from ert.simulator.forward_model_status import ForwardModelStatus
